+++ .idea/misc.xml
@@ -4,6 +4,8 @@
     <option name="filePathToZoomLevelMap">
       <map>
         <entry key="..\:/Users/AndroidDev/AppData/Local/Android/Sdk/platforms/android-32/data/res/layout/simple_spinner_item.xml" value="0.17552083333333332" />
+        <entry key="..\:/Users/HP/AndroidStudioProjects/RajVavahBioDataFetch/app/src/main/res/layout/activity_biodatadisplay.xml" value="0.25416666666666665" />
+        <entry key="..\:/Users/HP/AndroidStudioProjects/RajVavahBioDataFetch/app/src/main/res/layout/biodatadisplay_fragment.xml" value="0.25416666666666665" />
         <entry key="..\:/Users/HP/AndroidStudioProjects/RajVivahMaster/app/src/main/res/drawable/custom_rectangle.xml" value="0.195" />
         <entry key="..\:/Users/HP/AndroidStudioProjects/RajVivahMaster/app/src/main/res/drawable/icon_account.xml" value="0.205" />
         <entry key="..\:/Users/HP/AndroidStudioProjects/RajVivahMaster/app/src/main/res/drawable/icons_facebook.xml" value="0.195" />
