--- app/src/main/res/layout/activity_biodatadisplay.xml
@@ -1,35 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<androidx.core.widget.NestedScrollView
-    xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:id="@+id/idNestedSV"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    tools:context=".BiodatadisplayActivity">
-
-    <!--linear layout for displaying
-        our recycler view-->
-    <LinearLayout
-        android:layout_width="match_parent"
-        android:layout_height="match_parent"
-        android:orientation="vertical">
-
-        <!--recycler view for displaying our list of data
-            and we are making nested scroll for our
-            recycler view as false-->
-        <androidx.recyclerview.widget.RecyclerView
-            android:id="@+id/idRVCourses"
-            android:layout_width="match_parent"
-            android:layout_height="0dp"
-            android:layout_weight="1"
-            android:nestedScrollingEnabled="false" />
-
-        <!--we are adding progress bar
-            for the purpose of loading-->
-        <ProgressBar
-            android:id="@+id/idPBLoading"
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content" />
-    </LinearLayout>
-
-</androidx.core.widget.NestedScrollView>
